Changed size and max_size to size_t from int.

+++ IndexedContainer.h
@@ -14,7 +14,7 @@
 #endif
 
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 class IndexedContainer
 {
 public:
@@ -25,15 +25,15 @@ public:
   int add(const T &value);
   bool remove(const int index);
   bool indexHasValue(const int index);
-  int size();
-  int max_size();
+  size_t size();
+  size_t max_size();
   bool empty();
   bool full();
 
 private:
   T values_[MAX_SIZE];
   bool available_[MAX_SIZE];
-  int size_;
+  size_t size_;
 };
 #include "IndexedContainerDefinitions.h"
