+++ IndexedContainerDefinitions.h
@@ -9,22 +9,22 @@
 #define INDEXED_CONTAINER_DEFINITIONS_H
 
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 IndexedContainer<T,MAX_SIZE>::IndexedContainer()
 {
   clear();
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 T& IndexedContainer<T, MAX_SIZE>::operator[](const int index)
 {
-  if ((index >= 0) && (index < MAX_SIZE) && !available_[index])
+  if ((index >= 0) && (index < (int)MAX_SIZE) && !available_[index])
   {
     return values_[index];
   }
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 T& IndexedContainer<T, MAX_SIZE>::at(const int index)
 {
   if ((index >= 0) && (index < MAX_SIZE) && !available_[index])
@@ -33,24 +33,24 @@ T& IndexedContainer<T, MAX_SIZE>::at(const int index)
   }
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 void IndexedContainer<T, MAX_SIZE>::clear()
 {
   size_ = 0;
-  for (int index=0; index<MAX_SIZE; ++index)
+  for (size_t index=0; index<MAX_SIZE; ++index)
   {
     available_[index] = true;
   }
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 int IndexedContainer<T, MAX_SIZE>::add(const T &value)
 {
   if (size_ < MAX_SIZE)
   {
     bool found_available = false;
     int index = -1;
-    while (!found_available && (++index < MAX_SIZE))
+    while (!found_available && (++index < (int)MAX_SIZE))
     {
       found_available = available_[index];
     }
@@ -65,10 +65,10 @@ int IndexedContainer<T, MAX_SIZE>::add(const T &value)
   return -1;
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 bool IndexedContainer<T, MAX_SIZE>::remove(const int index)
 {
-  if ((index >= 0) && (index < MAX_SIZE) && (!available_[index]))
+  if ((index >= 0) && (index < (int)MAX_SIZE) && (!available_[index]))
   {
     available_[index] = true;
     --size_;
@@ -77,35 +77,35 @@ bool IndexedContainer<T, MAX_SIZE>::remove(const int index)
   return false;
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 bool IndexedContainer<T, MAX_SIZE>::indexHasValue(const int index)
 {
-  if ((index < 0) || (index >= MAX_SIZE))
+  if ((index < 0) || (index >= (int)MAX_SIZE))
   {
     return false;
   }
   return !available_[index];
 }
 
-template <typename T, int MAX_SIZE>
-int IndexedContainer<T, MAX_SIZE>::size()
+template <typename T, size_t MAX_SIZE>
+size_t IndexedContainer<T, MAX_SIZE>::size()
 {
   return size_;
 }
 
-template <typename T, int MAX_SIZE>
-int IndexedContainer<T, MAX_SIZE>::max_size()
+template <typename T, size_t MAX_SIZE>
+size_t IndexedContainer<T, MAX_SIZE>::max_size()
 {
   return MAX_SIZE;
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 bool IndexedContainer<T, MAX_SIZE>::empty()
 {
   return size_ == 0;
 }
 
-template <typename T, int MAX_SIZE>
+template <typename T, size_t MAX_SIZE>
 bool IndexedContainer<T, MAX_SIZE>::full()
 {
   return size_ == MAX_SIZE;
